backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.5.0
PyJWT>=2.0.0
passlib==1.7.4
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
